added ability to copy post text, title, and name. Also added reviewed changes
