+++ app/src/main/res/values/strings.xml
@@ -66,17 +66,17 @@
     <string name="comment_edit_type_your_comment">Type your comment</string>
     <string name="comment_mention_node_block">Block %1$s</string>
     <string name="comment_mention_node_copy_permalink">Copy Permalink</string>
-    <string name="comment_mention_node_copy_comment">Copy Comment</string>
+    <string name="comment_mention_node_copy_comment">Copy Comment Text</string>
     <string name="comment_mention_node_go_to">Go to %1$s</string>
     <string name="comment_mention_node_permalink_copied">Permalink Copied</string>
-    <string name="comment_mention_node_comment_copied">Comment Copied</string>
+    <string name="comment_mention_node_comment_copied">Comment Text Copied</string>
     <string name="comment_mention_node_report_comment">Report Comment</string>
     <string name="comment_mention_node_view_source">View Source</string>
     <string name="comment_next_parent">Next Parent Comment</string>
     <string name="comment_previous_parent">Previous Parent Comment</string>
     <string name="comment_node_block">Block %1$s</string>
     <string name="comment_node_copy_permalink">Copy Permalink</string>
-    <string name="comment_node_copy_comment">Copy Comment</string>
+    <string name="comment_node_copy_comment">Copy Comment Text</string>
     <string name="comment_node_delete">Delete</string>
     <string name="comment_node_edit">Edit</string>
     <string name="comment_node_go_to">Go to %1$s</string>
@@ -84,7 +84,7 @@
     <string name="comment_node_in">"in "</string>
     <string name="comment_node_more_replies">%1$s more replies</string>
     <string name="comment_node_permalink_copied">Permalink Copied</string>
-    <string name="comment_node_comment_copied">Comment Copied</string>
+    <string name="comment_node_comment_copied">Comment Text Copied</string>
     <string name="comment_node_report_comment">Report Comment</string>
     <string name="comment_node_restore">Restore</string>
     <string name="comment_node_view_context">View Context</string>
@@ -262,14 +262,20 @@
     <string name="post_listing_comments_count">%1$s comments</string>
     <string name="post_listing_copy_link">Copy link</string>
     <string name="post_listing_copy_permalink">Copy Permalink</string>
-    <string name="post_listing_copy_thumbnail_url">Copy Thumbnail Url</string>
+    <string name="post_listing_copy_thumbnail_link">Copy Thumbnail Link</string>
+    <string name="post_listing_copy_title">Copy Post Title</string>
+    <string name="post_listing_copy_name">Copy Post Name</string>
+    <string name="post_listing_copy_text">Copy Post Text</string>
     <string name="post_listing_delete">Delete</string>
     <string name="post_listing_edit">Edit</string>
     <string name="post_listing_go_to">Go to %1$s</string>
     <string name="post_listing_link_copied">Link Copied</string>
     <string name="post_listing_new">( %1$s new )</string>
     <string name="post_listing_permalink_copied">Permalink Copied</string>
-    <string name="post_listing_thumbnail_url_copied">Thumbnail Url Copied</string>
+    <string name="post_listing_thumbnail_link_copied">Thumbnail Link Copied</string>
+    <string name="post_listing_title_copied">Title Copied</string>
+    <string name="post_listing_name_copied">Name Copied</string>
+    <string name="post_listing_text_copied">Text Copied</string>
     <string name="post_listing_report_post">Report Post</string>
     <string name="post_listing_restore">Restore</string>
     <string name="privateMessage_reply">Reply to message</string>
@@ -322,5 +328,5 @@
     <string name="inbox_activity_mentions">Mentions</string>
     <string name="inbox_activity_messages">Messages</string>
     <string name="bookmarks_activity_saved">Saved</string>
-    <string name="generic_error">An error has occurred.</string>
+    <string name="generic_error">An error has occurred</string>
 </resources>
